`BaseMigrateController` extracted
